completed assignment 16 - added refresh functionality
